rename component name; fix class name; remove useless html
